refactor: split app-global control arms into ControlServer+AppCommands
ControlServer+SessionActions.swift held both the per-session arms and the
app-global ones, pushing past the file-size budget with the session.restore
arm still to come.

Move the app-global arms — the tree projection, sidebar visibility/mode and
expand/collapse, keymap and config reload, themes, and the quick terminal —
into a sibling ControlServer+AppCommands.swift, matching the existing
+SurfaceIO/+WindowCommands split. Pure code motion apart from dropping the
thin wrappers the sidebar arms had accumulated: the implementations now carry
the ControlActions names directly.
