+++ ARCHITECTURE.md
@@ -34,7 +34,7 @@ The programmatic control channel (`agtermctl` over a unix socket) follows the sa
 
 - **`agtermCore`** holds the wire protocol (`ControlProtocol.swift`: `Command`, `ControlArgs`, `ControlRequest`/`ControlResponse`, the tree node types) and the pure resolvers (`ControlResolve.swift`: the target resolver and the socket-path resolver). The `agtermCore` *library* target stays dependency-free.
 - **`agtermctlKit` + `agtermctl`** are separate targets in the same SwiftPM package. `agtermctlKit` is the testable library (the `swift-argument-parser` `ParsableCommand` tree plus the socket client); `agtermctl` is the thin executable. Only these CLI targets link `ArgumentParser`, so they build and test with `swift build`/`swift test` and never touch Xcode or GhosttyKit.
-- **`ControlServer`** (`agterm/Control/ControlServer.swift`) lives in the app target. It owns the POSIX unix socket, decodes each request, and dispatches it onto the existing `AppActions`/`AppStore` seam (and `GhosttySurfaceView.inject(text:)` for input). It is the only piece that touches the live surfaces, keeping the protocol and the CLI host-free. The flat command dispatch is split for size across sibling extension files (`ControlServer+SessionActions`/`+SurfaceIO`/`+WindowCommands`), and target resolution (frontmost/cross-window store scoping, id/prefix matching, the pinned wire-error strings) lives in a small injected `ControlTargetResolver`.
+- **`ControlServer`** (`agterm/Control/ControlServer.swift`) lives in the app target. It owns the POSIX unix socket, decodes each request, and dispatches it onto the existing `AppActions`/`AppStore` seam (and `GhosttySurfaceView.inject(text:)` for input). It is the only piece that touches the live surfaces, keeping the protocol and the CLI host-free. The flat command dispatch is split for size across sibling extension files (`ControlServer+SessionActions`/`+SurfaceIO`/`+WindowCommands`/`+AppCommands`, the last holding the app-global arms — the tree projection, sidebar, keymap/config reload, themes, and the quick terminal), and target resolution (frontmost/cross-window store scoping, id/prefix matching, the pinned wire-error strings) lives in a small injected `ControlTargetResolver`.
 
 See `CLAUDE.md` for the socket lifecycle, addressing, command catalog, and the keep-in-sync convention.
